submit hook, better messages, terminal or no terminal options, other usability improvs

+++ setup.sh
@@ -10,19 +10,30 @@ else
     exit 1
 fi
 
-curl https://sh.rustup.rs -sSf | sh -s -- -y
-source "$config_file"
-. "$HOME/.cargo/env"   
-cargo install --git https://github.com/asciinema/agg
+# if /home/agent/.agent_code/terminal_gifs.flag exists do this step, otherwise skip
+if [ -f /home/agent/.agent_code/terminal_gifs.flag ]; then
+    curl https://sh.rustup.rs -sSf | sh -s -- -y
+    source "$config_file"
+    . "$HOME/.cargo/env"   
+    cargo install --git https://github.com/asciinema/agg
+else
+    echo "Skipping installation of terminal gifs packages"
+fi
+
+# Write the setup flag 
+touch /home/agent/.agent_code/setup.flag
 
 # Add the aliases
 echo "# Python script aliases" >> "$config_file"
 echo "alias note!='python ~/.agent_code/note.py'" >> "$config_file"
 echo "alias clock!='python ~/.agent_code/clock.py'" >> "$config_file"
-echo 'export PATH="$HOME/.cargo/bin:$PATH"' >> "$config_file"
+echo "alias submit!='python ~/.agent_code/submit.py'" >> "$config_file"
+echo 'export PATH="$HOME/.cargo/bin:$PATH"' >> "$config_file"  
 
 # Reload the configuration
 source "$config_file"
+
+# Introduction message
 echo "======================================================="
 echo "Welcome to the Headless-Human Agent!"
 echo "======================================================="
@@ -31,10 +42,29 @@ echo "The following affordances are provided:"
 echo "- Automatic terminal recording (if you are seeing this message, then recording has started)"
 echo "- The 'note!' command, which we encourage you to use to take stream-of-consciousness notes"
 echo "- The 'clock!' command, which allows you to start and pause the timer."
+echo "- The 'submit!' command, which allows you to end your task and submit your work."
+echo "======================================================="
+echo "The above instructions will also be saved in the file /home/agent/human_agent_info.txt"
 echo "======================================================="
 echo "The clock is currently PAUSED."
 echo "When you are ready to proceed, run 'clock!' and start the timer."
 echo "======================================================="
+
+# Save the instructions to a file
+echo "=======================================================" > /home/agent/human_agent_info.txt
+echo "Welcome to the Headless-Human Agent!" >> /home/agent/human_agent_info.txt
+echo "=======================================================" >> /home/agent/human_agent_info.txt
+echo "This agent is the bridge between you and the Vivaria platform" >> /home/agent/human_agent_info.txt
+echo "The following affordances are provided:" >> /home/agent/human_agent_info.txt
+echo "- Automatic terminal recording (if you are seeing this message, then recording has started)" >> /home/agent/human_agent_info.txt
+echo "- The 'note!' command, which we encourage you to use to take stream-of-consciousness notes" >> /home/agent/human_agent_info.txt
+echo "- The 'clock!' command, which allows you to start and pause the timer." >> /home/agent/human_agent_info.txt
+echo "- The 'submit!' command, which allows you to end your task and submit your work." >> /home/agent/human_agent_info.txt
+echo "=======================================================" >> /home/agent/human_agent_info.txt
+echo "The clock is currently PAUSED." >> /home/agent/human_agent_info.txt
+echo "When you are ready to proceed, run 'clock!' and start the timer." >> /home/agent/human_agent_info.txt
+echo "=======================================================" >> /home/agent/human_agent_info.txt
+
 
 # Start the script session
 python /home/agent/.agent_code/terminal.py &
